Only age files with the same name
File detail aging was not neither looking if the directory is part of
the current category nor was it looking if the file name is the same.

With this change only file details are aged that are actually part of
the current category and if they have the same file name.

Signed-off-by: Adrian Reber <adrian@lisas.de>