Return an error if the checksum URL returns an error
To create checksums for the database scan-primary-mirror uses a HTTP
connection to download the target file. scan-primary-mirror does not
check the return code, so if the corresponding URLs returns a 404
scan-primary-mirror will happily create the checksum from the 404
document. This adds an additional step to check if the status is 'OK'.

Signed-off-by: Adrian Reber <adrian@lisas.de>